aufs: for 3.6-rc1, proc_map.patch
Here is a patch for 3.6-rc1. Still testing if that kernel compiles for
me ...

(jro: Since aufs3-proc_map.patch is generated automatically, this commit
will be overwritten. But it is worth to commit the patch from
users. Additionally it will be helpful for future update on my side at
release-time.)

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>